[css-scroll-snap] when navigating to a fragment... 他
Remove reference to 'point' being at-risk, as it was already punted to
the next level.

Per WG resolution, when navigating to a fragment: UAs must use snap
positions if snapping is on, may use them if snapping is off, should
pay attention to scroll-snap-margin regardless
https://github.com/w3c/csswg-drafts/commit/17f6d8d8dc1e0f7f20db04db2a8cb
50452177337

Use <length-percentage>, fix linking errors.
https://github.com/w3c/csswg-drafts/commit/091f30ab9da51e66da4a676298931
3d24480c48e