[selection-api] Define Selection.prototype.modify 他
Define Selection.prototype.modify
https://github.com/w3c/selection-api/commit/12d6a89e81e8646de080d580d62250b631e3b939

Remove the stray sentence that editing hosts outside a document cannot have a selection
https://github.com/w3c/selection-api/commit/186ddc8a6baf17c393cc38622afbe69d9f6d9a05

Clarify that the user interaction task source is used for selectionchange event
https://github.com/w3c/selection-api/commit/5fea97d3b53e86d437ccff5a1476621556cf5fdb

Add security and privacy considerations.
https://github.com/w3c/selection-api/commit/7a78eb16d241cbcfa18608453ae72e2c12b2d4eb

Clarify that the direction of selection is preserved when scripts mutates its range.
https://github.com/w3c/selection-api/commit/fff9e059fcb5fa9f9c49bfa9c76219a21846068e

Align description of events with DOM suggested phrasing
https://github.com/w3c/selection-api/commit/677612d5a2914f2e7d84f7492a82039d1b25d731

Fix the definition of containsNode() method when allowPartialContainment is true.
https://github.com/w3c/selection-api/commit/721b71658fa6f870ee5a8934d68744207638b9de

Closes the issue 126.
https://github.com/w3c/selection-api/commit/38f7f8c17dde1b0177b10264d156053b69badf3e

Remove the outdated note to address the issue 128
https://github.com/w3c/selection-api/commit/be0c73f215955fe18884188e0412f3dc68e2ff26