[css-values-5] 各種更新
Correct the translation of a progress between the two color-mix and cross-fade variants.
https://github.com/w3c/csswg-drafts/commit/19370ef74283428468e0af9e50ee47d09e07bdd8

Make progress functions all return numbers *with* a percent hint if needed. Allow <progress> to accept that. Mix in the <progress> percent hint into the value of calc-mix().
https://github.com/w3c/csswg-drafts/commit/9f7766f55165da2639ab8d9504b935cbefd2f851

Set the types correctly for progress fucntions and calc-mix().
https://github.com/w3c/csswg-drafts/commit/0d0efea0e2bdce3d5fde9bb1c433a951eaed8790