[css-anchor-position] 各種更新
Define that env() flips its variable according to the try tactic.
https://github.com/w3c/csswg-drafts/commit/e143edf116418e86e6c940061515d856a70a8d2d

Added the standard Value Definitions section, so it could hold the note about the flat tree.
https://github.com/w3c/csswg-drafts/commit/dcee5924b5f4db542afb4ead00d7f16acba609f0

Make anchor-size usable in the other accepted @position-try properties.
https://github.com/w3c/csswg-drafts/commit/1057a8d490bad708dacd4a02b51e134489fa0caa

Remove the logical props; their grammars are already defined in reference to one of the physical props.
https://github.com/w3c/csswg-drafts/commit/cc0ecf87c2a3ee4fc2344603f5e91f9f831a0cfa

Add propdef table for anchor() and anchor-size() so they show up in webref.
https://github.com/w3c/csswg-drafts/commit/0e10677089431ffbb6e9c24ab330feb178aa9349

Rewrite the ordering conditions on 'acceptable anchor element' to be more explicit and hopefully clearer.
https://github.com/w3c/csswg-drafts/commit/fe8e40d37faae8323d519f8d2b0a9dec7694e761