[css-easing-2] Rewrite the linear()...
Rewrite the linear() output algo to handle a single point (and correctly handle a number of edge cases).
https://github.com/w3c/csswg-drafts/commit/aaad9c74937acfc63f9f67ba24b530d0b5992a22

editorial: Some markup nits, and add a graph to linear like the others have.
https://github.com/w3c/csswg-drafts/commit/a8f5f8c980232e483ab34a6fa8bb73c118edf247

Rewrite serialization for linear() to be simpler. Make the linear/linear() grouping the same as the other keywords+functions.
https://github.com/w3c/csswg-drafts/commit/9e96e70a65c22fa5215a001e3cd48ae4920f1d8c