[css-values-5] Replace the notion of arb-sub functions... 他
editorial: Add an example of style() being cyclic.
https://github.com/w3c/csswg-drafts/commit/148dda6cc5582be6a5dac249b1381aaec5c68879

Specify that style queries in if() introduce a dependency.
https://github.com/w3c/csswg-drafts/commit/c7235bb462741475919566b9e1523f00d20258df

Replace the notion of arb-sub functions creating a dependency graph with them tracking a stack of dependencies as they evaluate. (Should be editorial, but it's a deep rewrite.) Give arb-sub functions a 2-stage grammar, allowing them to control when their bits are evaluated. Add spread syntax to arb-sub functions.
https://github.com/w3c/csswg-drafts/commit/f7a077668ea06a0509203c65d4d3cd5ef25ff458

＋ 編集中の更新を誤って commit した内容：
https://github.com/triple-underscore/triple-underscore.github.io/commit/70a91e85c48c208b00bbaa0e8b9f76cb360b5e6e