[css-color-4] 各種更新
Clarify that inside the color property, it is the resolved inherited value (not the raw inherited value) that is used.
https://github.com/w3c/csswg-drafts/commit/bad5b16e100bacff172608728e85a553d23453c2

Listed categories of colors, such as those that resolve to sRGB or support legacy color syntax.
https://github.com/w3c/csswg-drafts/commit/7ed51a851d232e520e3c240d09ce3b3d0c71a132

Added hue normalization examples
https://github.com/w3c/csswg-drafts/commit/bf78a61893a3f113d1967998cdf7ef2cadd13fb0

Corrected table of analogous components, alpha was missing, but described as analogous in prose
https://github.com/w3c/csswg-drafts/commit/b530ae0dba0d794e86676e1c0f9f8f4f1b5fd5df